Fix Size for maps.
This was terribly broken, but we got lucky for small map entries.
The old code was trying to compute the size of the tag codes for
the key and value, but that's the responsibility of the sizer for
the particular key and value. What size_new_map needs to account for
is the length of the tag code and length varint of the map entry itself.

Fixes #21.