Print out malformed generated code in a useful way.
While working on the code generator, I occasionally cause it to make
malformed Go code. It can often be very hard to trace the problem,
since the only error message that is printed is the compiler-style
error from go/parser; without the corresponding source code it is
usually a useless message. So, if we generate bad source code, print
out that source code at the same time.