Update README to be Markdown formatted README.md.
I missed the installation of the C++ protocol buffer compiler step when
I first started. This is because with the plain-text formatting it was
hard to read, so I skipped to the line "to install this code." By adding
headings, it should make this mistake less common.

I also added quotes to enable
[syntax
highlighting](https://help.github.com/articles/github-flavored-markdown/#syntax-highlighting),
but tried to keep edits to a minimum.