Add headings to make clear how to get started.
With the previous formatting it was easy to miss the step to install the
C++ version of the protocol buffer compilers. Now it will be easy to
find out what to do to start using protocol buffers with Go.

I rename the file to README.md, because I found the plain-text rendering
on GitHub to be especially difficult to read, but I kept the formatting
changes to a minimum.