Add headings to make clear how to get started.
Change the installation steps to be formatted as a list. This way it is
clear what all the steps are.

With the previous formatting it was easy to miss the step to install the
C++ version of the protocol buffer compilers. Now it will be easy to
find out what to do to start using protocol buffers with Go.