Add TimestampNow and TimestampFormat funcs
This adds two timestamp functions to the pytpe library. TimestampNow()
returns a protobuf Timestamp with the current time. TimestampFormat()
takes a Go time format string and formats a protobuf Timestamp
accordingly. With TimestampFormat here, TimestampString() was refactored
to use that function.

Additionally, one documentation fix was made to the Timestamp()
function.