remove most of the structPointer abstraction.
It was pointless for me (it exists to support the reflection-based code
meant for 'safe' enviroments which can't use the unsafe package.

And reading it was making my brain hurt. Now it still hursts (too much
relection-based code always does for me) but at least the pointer math
and the casts are plain and obvious.

While I am at it, hoist the isMarshaler tests out of the loops.