fix name collision in protoc between field and type
turns out protoc 3.0.2 does not keep separate namespaces for types and
message fields. So

 message Msg {
   message Y { ... }
   Y Y = 1;
 }

fails to compile because the name "Y" is used twice.

One solution I tried was to rename the type to Y_msg. However after
looking at the output it didn't look like what a human would do.
Looking at other .proto files I realized a human would not have this
problem because by and large the type names start with uppercase while
the field names are lowercased. So I did the same. Now the go

 type Msg struct {
   SecondMsg struct { ... } `protobuf:"bytes,1"`
 }

turns into

 message Msg {
  message SecondMsg { ... }
  SecondMsg second_msg = 1;
 }

which is useable. The protobuf field name can be overridden with
a 'name=xxx' tag. As in `protobuf:"bytes,1,name=xxx"`. And some
other heuristics exist. See comments in source code.

Note: another possibility woudl be to only lowercase the 1st
letter of the field. However that didn't look nice to me.