have AsProtobufFull() emit the types in sorted order
rather than in map iteration order (aka random). This makes
diff'ing successive outputs easy.

Now if only I could preserve the field's comments in the .go source
I'd be really happy. The only idea I have is to embed the comments
in field tags, which feels wrong. (Well the other is to read the
source AST and dig, but that's difficult)