cleanup names. field->offset, and structPointer->unsafe.Pointer
that removes a number of casts and make the code much clearer.

(the whole structPointer thing was so that the package could have
two implementations. This one which uses unsafe.Pointer, and a slower one
which could run in restricted environments which used relflect's Set()
methods. We don't intend to run slow)