Fixed a small issue with the go generator's self-import check.
Previously the go generator only checked to ensure that original package name is not self-imported. When the  option is used, this check does not cover the option-defined package. This change adds in that check to ensure that the go generator doesn't self import in addition to avoiding prepending package names to types in the generated go code for those same option-defined packages