fix decoding of zigzag32 negative values once cast to a 64-bit 'int' type
the bug dates back to a [mis]feature of the upstream google code: since
their decoder never puts the result of a DecodeZigzag32() in anything
but an 'int32' type it didnt matter what the upper 32 bits of the result
where. They had them hardcoded to 0. Aka the signed 32-bit result was
embedded in the lower 32 bits of the unsigned 64-bit result.

That breaks in my code, where I might cast the result to an 'int' on
targets where 'int' is 64-bits. The fix is to return a signed 64-bit
result. Then everyone is happy. That does require an adjustment to
all_test.go, since the raw 64-but result changed.