proto: reject invalid UTF-8 in strings
The proto2 and proto3 specifications explicitly say that strings must be
composed of valid UTF-8 characters. Thus, report a marshal/unmarshal error
if any string fields are invalid.

The C++ implementation prints an error anytime invalid UTF-8 strings
is present during parsing or serializing. For proto3, it explicitly
errors out when parsing an invalid string.

In Go, we'll attempt to apply this check on proto2 and proto3,
but may relax that constraint depending on the extent of problems.