Rename generated Marshal and Unmarshal protobuf method
As part of the recent protobuf optimization work, we started generating
Unmarshal and Marshal methods on generated messages. These were done to
support the table-driven implementation within the proto package.

We are renaming Marshal as XXX_Marshal and Unmarshal as XXX_Unmarshal
to make it clear that users are not supposed to call these directly.
There are several reasons for the rename:
* The Marshal method assumes that deterministic is the only option we
would ever support for protos, and is not very forward compatible.
* The presence of a deterministic boolean is confusing for users,
where many set it without considering whether it is necessary.
* The Unmarshal method has a slightly different semantic than the previously
documented proto.Unmarshaler interface. The documented Unmarshal specifies
that Reset is called, while the new method does no such thing. The semantic
difference warrants a rename of the method.
* Some users in the Go community depend on these methods not being
generated by default.

Fixes #530