catch cases when people use the 'bytes' wiretype for float64 and other types
which don't support this encoding. Basically any combination which
results in an .enc function which uses .valEnc cannot be used with
a wiretype which leaves valEnc nil (which is basically 'bytes' at the
moment). Since the decoders are symmetric with the encoder, checking
the encoder is sufficient.

A new unit test purposely exercises this.