rework properties in preparation for handling embedded fields
Now StructProperties.props is sorted by tag id. That eliminates
the indirection through StructProperties.order, and .order entirely.

No code used the fact that .props was indexed by reflection field number
and when we embed fields (whichi is coming next) that no longer made any sense.

The check for duplicate IDs is now done after the sorting by
tag id, which means it can just be a linear scan (it was using
a hash map as a set)