support for embedded fields in structs
This commit adds support for a magic tag `protobuf:"embedded"`
If it is found on an embedded struct field then the fields
of the embedded type are merged into those of the outer struct,
just like Go and the stdlib json.Marshaler do.

And if the embedded field has a regular tag like `protobuf:"bytes,1"`
then the field is treated as a regular named field, and its value is
wrapped in a "bytes,1" tag.