Refactor protoc-gen-go generator - split up generateMessage
Split actual text generation passes of generateMessage out by
first building a data structure suitable for code generation
and then generate the output for each field in separate methods.

Partially incorrect and unneccessary usage of g.In() and g.Out()
has als been removed throughout the code generation as the output
is properly formatted at the end of code generation anyway.

The functionality is unchanged except for the order of oneof
methods in the output file.