protoc-gen-go: move package comments to before the package decl
We intentionally don't attach comments from the .proto source file to
the Go package declaration (since this produces strange output when
multiple .proto files in the same Go package all have comments), but
putting the comment above the package decl makes it easier to find and
keeps it from being obscured by the imports.

Also change the package comment to be generated with PrintComments
(producing // comments rather than /* */ comments) to save a few lines
of code and be internally consistent.